+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+### Added
+
+- Data types now implement common traits (`Debug`, `Clone`, `Eq`, `PartialEq`,
+`Ord`, `PartialOrd`, `Hash`, `Default` and `Copy` ) where applicable
+
 ### Changed
 
 -   Remove `Qbe` prefix from data structures. `QbeValue` becomes `qbe::Value`